# api/internal/handler/cycle/cycle_handler.go
// Package cycle provides HTTP handlers for the evaluation lifecycle API.
// Handlers are thin: they validate input, call the service layer, and format
// responses. Business logic lives in the service layer.
package cycle

import (
	"encoding/json"
	"log"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/sed-evaluacion-desempeno/api/internal/middleware"
	pkgerrors "github.com/sed-evaluacion-desempeno/api/internal/pkg/errors"
	repoorg "github.com/sed-evaluacion-desempeno/api/internal/repository/org"
	svcgoal "github.com/sed-evaluacion-desempeno/api/internal/service/goal"

	"github.com/sed-evaluacion-desempeno/api/internal/auth"
	activitysvc "github.com/sed-evaluacion-desempeno/api/internal/service/activity"
	svc "github.com/sed-evaluacion-desempeno/api/internal/service/cycle"
)

// contextKey for handler-specific context values.
type contextKey struct{ name string }

var idempCtxKey = &contextKey{"idempotency-key"}

// generateTraceID generates a short trace ID for error responses.
func generateTraceID() string {
	id := uuid.New().String()
	return id[:8] + "-" + id[9:13]
}

// writeJSON writes a JSON response with the given status code.
func writeJSON(w http.ResponseWriter, status int, v interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(v); err != nil {
		log.Printf("handler: failed to encode JSON response: %v", err)
	}
}

// writeError writes a structured error response.
func writeError(w http.ResponseWriter, err error) {
	traceID := generateTraceID()
	status := pkgerrors.HTTPStatus(err)

	var de *pkgerrors.DomainError
	if pkgerrors.AsDomainError(err, &de) {
		writeJSON(w, status, pkgerrors.NewAPIErrorResponse(de, traceID))
		return
	}

	writeJSON(w, status, pkgerrors.NewAPIErrorResponse(
		pkgerrors.NewDomainError(pkgerrors.InvalidRequest, err.Error(), err),
		traceID,
	))
}

// CycleHandler holds HTTP handlers for cycle operations.
type CycleHandler struct {
	svc          svc.Service
	phaseService svc.PhaseService
	activitySvc  activitysvc.Service
	assignRepo   svcgoal.AssignmentRepository
	employeeRepo *repoorg.EmployeeRepo
}

// NewCycleHandler creates a new CycleHandler.
func NewCycleHandler(svc svc.Service, phaseService svc.PhaseService, activitySvc activitysvc.Service, assignRepo svcgoal.AssignmentRepository, employeeRepo *repoorg.EmployeeRepo) *CycleHandler {
	return &CycleHandler{
		svc:          svc,
		phaseService: phaseService,
		activitySvc:  activitySvc,
		assignRepo:   assignRepo,
		employeeRepo: employeeRepo,
	}
}

// ListCycles handles GET /api/v1/cycles.
func (h *CycleHandler) ListCycles(w http.ResponseWriter, r *http.Request) {
	orgID := r.URL.Query().Get("organization_id")
	if orgID == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"organization_id query parameter is required", nil))
		return
	}

	var year *int
	if y := r.URL.Query().Get("year"); y != "" {
		yv, err := strconv.Atoi(y)
		if err != nil {
			writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
				"year must be a valid integer", err))
			return
		}
		year = &yv
	}

	var currentPhase *string
	if cp := r.URL.Query().Get("current_phase"); cp != "" {
		currentPhase = &cp
	}

	cursor := r.URL.Query().Get("cursor")

	limit := 20
	if l := r.URL.Query().Get("limit"); l != "" {
		lv, err := strconv.Atoi(l)
		if err != nil {
			writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
				"limit must be a valid integer", err))
			return
		}
		if lv < 1 || lv > 100 {
			writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
				"limit must be between 1 and 100", nil))
			return
		}
		limit = lv
	}

	req := svc.ListCyclesRequest{
		OrganizationID: orgID,
		Year:           year,
		CurrentPhase:   currentPhase,
		Cursor:         cursor,
		Limit:          limit,
	}

	result, err := h.svc.ListCycles(r.Context(), req)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, result)
}

// CreateCycle handles POST /api/v1/cycles.
func (h *CycleHandler) CreateCycle(w http.ResponseWriter, r *http.Request) {
	var req svc.CreateCycleRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"invalid JSON body", err))
		return
	}

	// Extract idempotency key from context (injected by middleware)
	req.IdempotencyKey = middleware.IdempotencyKeyFromContext(r.Context())

	cycle, err := h.svc.CreateCycle(r.Context(), req)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusCreated, cycle)
}

// GetCycle handles GET /api/v1/cycles/{id}.
func (h *CycleHandler) GetCycle(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id is required", nil))
		return
	}

	if _, err := uuid.Parse(id); err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id must be a valid UUID v4", err))
		return
	}

	cycle, err := h.svc.GetCycle(r.Context(), id)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, cycle)
}

// GetCurrentCycle handles GET /api/v1/cycles/current.
// Resolves the cycle for the given year, falling back to the active
// (unfinished) cycle. Must be registered BEFORE /cycles/{id} in routes.
// Query: organization_id (required, UUID), year (optional, defaults to current year).
func (h *CycleHandler) GetCurrentCycle(w http.ResponseWriter, r *http.Request) {
	orgID := r.URL.Query().Get("organization_id")
	if orgID == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"organization_id query parameter is required", nil))
		return
	}

	if _, err := uuid.Parse(orgID); err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"organization_id must be a valid UUID v4", err))
		return
	}

	year := 0
	if y := r.URL.Query().Get("year"); y != "" {
		yv, err := strconv.Atoi(y)
		if err != nil {
			writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
				"year must be a valid integer", err))
			return
		}
		year = yv
	}

	cycle, err := h.svc.GetCurrentCycle(r.Context(), orgID, year)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, cycle)
}

// TransitionPhase handles PUT /api/v1/cycles/{id}/transition.
func (h *CycleHandler) TransitionPhase(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id is required", nil))
		return
	}

	if _, err := uuid.Parse(id); err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id must be a valid UUID v4", err))
		return
	}

	// Parse optional body
	var body struct {
		Trigger string `json:"trigger"`
		ToPhase string `json:"to_phase"`
		Reason  string `json:"reason"`
	}
	_ = json.NewDecoder(r.Body).Decode(&body)

	// Get expected version from context (injected by OptimisticLock middleware)
	expectedVersion := middleware.ExpectedVersionFromContext(r.Context())

	// Get idempotency key from context
	idempotencyKey := middleware.IdempotencyKeyFromContext(r.Context())

	req := svc.TransitionPhaseRequest{
		CycleID:         id,
		ExpectedVersion: expectedVersion,
		Trigger:         body.Trigger,
		ToPhase:         body.ToPhase,
		Reason:          body.Reason,
		IdempotencyKey:  idempotencyKey,
	}

	result, err := h.svc.TransitionPhase(r.Context(), req)
	if err != nil {
		writeError(w, err)
		return
	}

	// Log activity: transición de fase del ciclo
	if h.activitySvc != nil {
		if empID, ok := auth.GetEmployeeID(r.Context()); ok {
			metadata := map[string]interface{}{
				"cycle_id": id,
				"trigger":  body.Trigger,
			}
			_ = h.activitySvc.LogActivity(r.Context(), empID, "cycle_configured",
				"Configuraste el ciclo de evaluación", "Ciclos", metadata)
		}
	}

	writeJSON(w, http.StatusOK, result)
}

// RevertPhase handles POST /api/v1/cycles/{id}/revert.
// Accepts an optional explicit target phase via ?to_phase= query param or
// JSON body {"to_phase": "..."}. When to_phase is set, the request is
// delegated to TransitionPhase (bidirectional, RH trigger); otherwise the
// cycle moves one step back via RevertPhase.
func (h *CycleHandler) RevertPhase(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id is required", nil))
		return
	}

	if _, err := uuid.Parse(id); err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id must be a valid UUID v4", err))
		return
	}

	// Optional explicit target: query param takes precedence, else JSON body.
	toPhase := r.URL.Query().Get("to_phase")
	if toPhase == "" {
		var body struct {
			ToPhase string `json:"to_phase"`
		}
		_ = json.NewDecoder(r.Body).Decode(&body)
		toPhase = body.ToPhase
	}

	if toPhase != "" {
		idempotencyKey := middleware.IdempotencyKeyFromContext(r.Context())
		cycle, err := h.svc.GetCycle(r.Context(), id)
		if err != nil {
			writeError(w, err)
			return
		}
		result, err := h.svc.TransitionPhase(r.Context(), svc.TransitionPhaseRequest{
			CycleID:         id,
			ExpectedVersion: cycle.Version,
			Trigger:         "manual_rh",
			ToPhase:         toPhase,
			Reason:          "revert to " + toPhase,
			IdempotencyKey:  idempotencyKey,
		})
		if err != nil {
			writeError(w, err)
			return
		}
		writeJSON(w, http.StatusOK, result)
		return
	}

	result, err := h.svc.RevertPhase(r.Context(), id)
	if err != nil {
		writeError(w, err)
		return
	}

	writeJSON(w, http.StatusOK, result)
}

// GetPhaseDefinitions handles GET /api/v1/phases.
func (h *CycleHandler) GetPhaseDefinitions(w http.ResponseWriter, r *http.Request) {
	defs, etag, err := h.phaseService.GetPhaseDefinitions(r.Context())
	if err != nil {
		writeError(w, err)
		return
	}

	// Check If-None-Match
	if r.Header.Get("If-None-Match") == `"`+etag+`"` {
		w.WriteHeader(http.StatusNotModified)
		return
	}

	w.Header().Set("ETag", `"`+etag+`"`)
	w.Header().Set("Cache-Control", "max-age=3600")

	resp := map[string]interface{}{
		"data": defs,
	}
	writeJSON(w, http.StatusOK, resp)
}

// GetAvailableTransitions handles GET /api/v1/cycles/{id}/transitions.
func (h *CycleHandler) GetAvailableTransitions(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id is required", nil))
		return
	}

	if _, err := uuid.Parse(id); err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id must be a valid UUID v4", err))
		return
	}

	transitions, err := h.phaseService.GetAvailableTransitions(r.Context(), id)
	if err != nil {
		writeError(w, err)
		return
	}

	resp := map[string]interface{}{
		"data": transitions,
	}
	writeJSON(w, http.StatusOK, resp)
}

// AssignAllEmployees handles POST /api/v1/cycles/{id}/assign-all.
// Assigns all active employees to the cycle. Idempotent: skips employees
// that already have an assignment for this cycle.
func (h *CycleHandler) AssignAllEmployees(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if id == "" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id is required", nil))
		return
	}

	cycleID, err := uuid.Parse(id)
	if err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle id must be a valid UUID v4", err))
		return
	}

	// Get cycle to validate phase and extract org ID
	cycle, err := h.svc.GetCycle(r.Context(), id)
	if err != nil {
		writeError(w, err)
		return
	}

	if cycle.CurrentPhase != "asignacion" {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"cycle must be in asignacion phase to assign employees", nil))
		return
	}

	// List all active employees for this cycle's organization
	orgID, err := uuid.Parse(cycle.OrganizationID)
	if err != nil {
		writeError(w, pkgerrors.NewDomainError(pkgerrors.InvalidRequest,
			"invalid organization ID in cycle", err))
		return
	}

	isActive := true
	employees, err := h.employeeRepo.List(r.Context(), repoorg.EmployeeFilter{
		TreeID:   &orgID,
		IsActive: &isActive,
		Limit:    10000,
	})
	if err != nil {
		writeError(w, err)
		return
	}

	assigned := 0
	skipped := 0
	for _, emp := range employees {
		_, err := h.assignRepo.CreateAssignment(r.Context(), emp.ID, cycleID)
		if err != nil {
			skipped++
			continue
		}
		assigned++
	}

	// Log activity
	if h.activitySvc != nil {
		if empID, ok := auth.GetEmployeeID(r.Context()); ok {
			metadata := map[string]interface{}{
				"cycle_id": id,
				"assigned": assigned,
				"skipped":  skipped,
			}
			_ = h.activitySvc.LogActivity(r.Context(), empID, "cycle_configured",
				"Asignaste empleados al ciclo de evaluación", "Ciclos", metadata)
		}
	}

	writeJSON(w, http.StatusOK, map[string]interface{}{
		"assigned": assigned,
		"skipped":  skipped,
		"total":    len(employees),
	})
}