# api/openapi/goals-api.yaml
openapi: 3.1.0
info:
  title: SED Goals API
  version: '1.0.0'
  description: Core employee-facing API for goals, categories, KPIs, weight validation, and assignments.
servers:
  - url: /api/v1
paths:
...
  /employees/{empId}/assignments:
    get:
      operationId: getAssignment
      summary: Get the employee's current goal assignment
      parameters:
        - $ref: '#/components/parameters/EmpId'
      responses:
        '200':
          description: Assignment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
        '404':
          $ref: '#/components/responses/GoalNotFound'
    post:
      operationId: createAssignment
      summary: Create a goal assignment for the employee
      parameters:
        - $ref: '#/components/parameters/EmpId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '201':
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
        '403':
          $ref: '#/components/responses/PhaseRestricted'

  # ============================================================================
  # Scoring
  # ============================================================================
...
components:
  parameters:
    EmpId:
      name: empId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: string
            trace_id:
              type: string

    # ============================================================================
    # Category Schemas
    # ============================================================================
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        weight:
          type: number
        goals:
          type: array
          items:
            $ref: '#/components/schemas/GoalResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GoalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        weight:
          type: number
        target_value:
          type: number
        current_value:
          type: number
        direction:
          type: string
          enum: [ascendente, descendente]
        baseline_value:
          type: number
          nullable: true
        progress_percent:
          type: number
        state:
          type: string
          enum: [borrador, fijada, en_seguimiento, evaluada, cerrada]
        version:
          type: integer
        kpis:
          type: array
          items:
            $ref: '#/components/schemas/KpiResponse'
        pending_proposal:
          $ref: '#/components/schemas/GoalProposalResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GoalProposalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        requested_by:
          type: string
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        weight:
          type: number
        target_value:
          type: number
        direction:
          type: string
          enum: [ascendente, descendente]
        baseline_value:
          type: number
          nullable: true
        kpi_ids:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
        reviewed_by:
          type: string
          nullable: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    KpiResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        unit:
          type: string
        description:
          type: string
        direction:
          type: string
          enum: [ascendente, descendente]
        org_node_id:
          type: string
          format: uuid
          nullable: true
          description: "Org node (department) that owns the KPI; null means a global KPI visible across the organization."
        current_value:
          type: number
          nullable: true
        progress_percent:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AssignmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        cycle_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [borrador, enviada]
          description: Estado actual de la asignación de metas
        submitted_at:
          type: string
          format: date-time
          nullable: true
          description: Fecha y hora en que la asignación fue enviada formalmente
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        global_goals:
          type: array
          items:
            $ref: '#/components/schemas/AssignedGoalResponse'
        shared_goals:
          type: array
          items:
            $ref: '#/components/schemas/AssignedGoalResponse'
        created_at:
          type: string
          format: date-time
    CreateAssignmentRequest:
      type: object
      required: [cycle_id]
      properties:
        cycle_id:
          type: string
          format: uuid
    AssignedGoalResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        description: {type: string}
        unit: {type: string}
        direction: {type: string}
        goal_kind: {type: string}
        weight: {type: number}
        target_value: {type: number}
        baseline_value: {type: number, nullable: true}
        current_value: {type: number}
        progress_percent: {type: number}
        state: {type: string}
        source: {type: string, enum: [global, shared]}
...
  responses:
    PhaseRestricted:
      description: Operation not allowed in current cycle phase
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    GoalNotFound:
      description: Goal not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...